Move xds protos to testing/
