+++ build.yaml
@@ -4755,7 +4755,7 @@ targets:
   build: test
   language: c++
   src:
-  - src/proto/grpc/lb/v2/orca_load_report_for_test.proto
+  - src/proto/grpc/testing/xds/orca_load_report_for_test.proto
   - test/cpp/end2end/client_lb_end2end_test.cc
   deps:
   - grpc++_test_util
@@ -6028,8 +6028,8 @@ targets:
   build: test
   language: c++
   src:
-  - src/proto/grpc/lb/v2/eds_for_test.proto
-  - src/proto/grpc/lb/v2/lrs_for_test.proto
+  - src/proto/grpc/testing/xds/eds_for_test.proto
+  - src/proto/grpc/testing/xds/lrs_for_test.proto
   - test/cpp/end2end/xds_end2end_test.cc
   deps:
   - grpc++_test_util
