+++ test/cpp/end2end/client_lb_end2end_test.cc
@@ -53,8 +53,8 @@
 #include "src/cpp/client/secure_credentials.h"
 #include "src/cpp/server/secure_server_credentials.h"
 
-#include "src/proto/grpc/lb/v2/orca_load_report_for_test.pb.h"
 #include "src/proto/grpc/testing/echo.grpc.pb.h"
+#include "src/proto/grpc/testing/xds/orca_load_report_for_test.pb.h"
 #include "test/core/util/port.h"
 #include "test/core/util/test_config.h"
 #include "test/core/util/test_lb_policies.h"
